mini-services/ml-service/models/signal_classifier.py
"""
Signal Classifier Model
Gradient Boosting classifier for buy/sell signal prediction.
"""

import numpy as np
from typing import Dict, Any, Optional
import logging

logger = logging.getLogger(__name__)

try:
    from sklearn.ensemble import GradientBoostingClassifier, RandomForestClassifier
    from sklearn.model_selection import cross_val_score
    from sklearn.preprocessing import StandardScaler
    SKLEARN_AVAILABLE = True
except ImportError:
    SKLEARN_AVAILABLE = False
    logger.warning("scikit-learn not available, using mock implementation")


class SignalClassifierModel:
    """
    Gradient Boosting classifier for trading signals.
    
    Features:
    - Multi-class classification (BUY, SELL, HOLD)
    - Probability calibration
    - Feature importance analysis
    - Cross-validation
    """
    
    SIGNAL_TYPES = ['HOLD', 'BUY', 'SELL']
    
    def __init__(
        self,
        n_estimators: int = 200,
        max_depth: int = 6,
        learning_rate: float = 0.1,
        min_samples_split: int = 10,
        min_samples_leaf: int = 5,
    ):
        self.n_estimators = n_estimators
        self.max_depth = max_depth
        self.learning_rate = learning_rate
        self.min_samples_split = min_samples_split
        self.min_samples_leaf = min_samples_leaf
        
        self.model = None
        self.scaler = None
        self.is_trained = False
        self.feature_importance = None
        self.cv_scores = None
        
        if SKLEARN_AVAILABLE:
            self.model = GradientBoostingClassifier(
                n_estimators=n_estimators,
                max_depth=max_depth,
                learning_rate=learning_rate,
                min_samples_split=min_samples_split,
                min_samples_leaf=min_samples_leaf,
                random_state=42,
            )
            self.scaler = StandardScaler()
    
    def train(
        self,
        X: np.ndarray,
        y: np.ndarray,
        cross_validate: bool = True
    ) -> Dict[str, Any]:
        """
        Train the classifier.
        
        Args:
            X: Feature matrix (samples, features)
            y: Labels (samples,) - 0=HOLD, 1=BUY, 2=SELL
            cross_validate: Whether to perform cross-validation
            
        Returns:
            Training metrics
        """
        if not SKLEARN_AVAILABLE or self.model is None:
            self.is_trained = True
            return {"cv_scores_mean": 0.5, "cv_scores_std": 0.1}
        
        # Scale features
        X_scaled = self.scaler.fit_transform(X)
        
        # Cross-validation
        if cross_validate:
            cv_scores = cross_val_score(self.model, X_scaled, y, cv=5)
            self.cv_scores = cv_scores
        
        # Fit final model
        self.model.fit(X_scaled, y)
        self.feature_importance = self.model.feature_importances_
        self.is_trained = True
        
        return {
            "cv_scores_mean": float(self.cv_scores.mean()) if self.cv_scores is not None else None,
            "cv_scores_std": float(self.cv_scores.std()) if self.cv_scores is not None else None,
            "feature_importance": self.feature_importance.tolist() if self.feature_importance is not None else None,
        }
    
    def predict(self, X: np.ndarray) -> np.ndarray:
        """
        Predict signal class.
        
        Args:
            X: Feature matrix
            
        Returns:
            Predicted classes (0=HOLD, 1=BUY, 2=SELL)
        """
        if not SKLEARN_AVAILABLE or self.model is None:
            return np.zeros(X.shape[0], dtype=int)
        
        X_scaled = self.scaler.transform(X)
        return self.model.predict(X_scaled)
    
    def predict_proba(self, X: np.ndarray) -> np.ndarray:
        """
        Predict class probabilities.
        
        Args:
            X: Feature matrix
            
        Returns:
            Probabilities (samples, 3) for [HOLD, BUY, SELL]
        """
        if not SKLEARN_AVAILABLE or self.model is None:
            probs = np.random.rand(X.shape[0], 3)
            return probs / probs.sum(axis=1, keepdims=True)
        
        X_scaled = self.scaler.transform(X)
        return self.model.predict_proba(X_scaled)
    
    def predict_signal(self, X: np.ndarray) -> Dict[str, Any]:
        """
        Predict with full signal information.
        
        Returns:
            Dict with signal, confidence, and probabilities
        """
        pred = self.predict(X)
        proba = self.predict_proba(X)
        
        signals = []
        for i in range(len(pred)):
            signals.append({
                "signal": self.SIGNAL_TYPES[pred[i]],
                "confidence": float(proba[i].max()),
                "probabilities": {
                    self.SIGNAL_TYPES[j]: float(proba[i, j])
                    for j in range(3)
                }
            })
        
        return signals
    
    def get_feature_importance(self) -> Optional[np.ndarray]:
        """Get feature importance"""
        return self.feature_importance
    
    def save(self, path: str) -> bool:
        """Save model to disk"""
        try:
            import joblib
            joblib.dump(self.model, f"{path}/signal_classifier.pkl")
            joblib.dump(self.scaler, f"{path}/signal_scaler.pkl")
            return True
        except Exception as e:
            logger.error(f"Failed to save model: {e}")
            return False
    
    def load(self, path: str) -> bool:
        """Load model from disk"""
        try:
            import joblib
            self.model = joblib.load(f"{path}/signal_classifier.pkl")
            self.scaler = joblib.load(f"{path}/signal_scaler.pkl")
            self.is_trained = True
            return True
        except Exception as e:
            logger.error(f"Failed to load model: {e}")
            return False
    
    def get_metrics(self) -> Dict[str, Any]:
        """Get model metrics"""
        return {
            "is_trained": self.is_trained,
            "n_estimators": self.n_estimators,
            "max_depth": self.max_depth,
            "learning_rate": self.learning_rate,
            "cv_scores_mean": float(self.cv_scores.mean()) if self.cv_scores is not None else None,
            "cv_scores_std": float(self.cv_scores.std()) if self.cv_scores is not None else None,
        }
