mini-services/ml-service/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
websockets>=12.0
numpy==1.26.3
pandas==2.2.0
scikit-learn==1.4.0
tensorflow==2.15.0
keras==2.15.0
h5py==3.10.0
joblib==1.3.2
pydantic==2.6.0
pyyaml==6.0.1
aiohttp==3.9.1
python-multipart==0.0.6
hmmlearn==0.3.0
starlette>=0.35.0
